requirements.txt
APScheduler==3.5.1
GroupyAPI==0.8.1
pytz==2018.3
requests==2.9.1
six==1.11.0
tzlocal==1.5.1
